Move 'value' to private folder
